requirements.txt
Django==5.1.7
python_jose==3.5.0
requests==2.32.5
Pillow
django-tinymce4==3.7.1
mysqlclient
